Create code for the game "Guess the number"
This code first sets a random number between 1 and 100 and then prompts the user to guess it. Each attempt of the user recently, and he is prompted whether his number is greater or less than the guessed one. When the number is guessed, the game finishes and reports how successful it was.